+++ src/credits/controllers/credits.controller.ts
@@ -2,16 +2,18 @@ import {
   UseFilters,
   Controller,
   Get,
-  Query,
   Req,
   Param,
   Post,
   Body,
+  UsePipes,
+  ValidationPipe,
+  Query,
 } from '@nestjs/common';
 import {
   ApiBearerAuth,
-  ApiExcludeEndpoint,
   ApiOkResponse,
+  ApiParam,
   ApiQuery,
   ApiTags,
 } from '@nestjs/swagger';
@@ -20,20 +22,57 @@ import { CreditService } from '../services/credits.service';
 import {
   CreditRequestDto,
   CreditResponseDto,
-  GetCreditsDto,
+  CreditPlanResponseDto,
+  GetCreditsQueryDto,
 } from '../dtos/credits.dto';
+import { CreditSourceEnum } from '../schemas/credit.schema';
 
 @UseFilters(AllExceptionsFilter)
 @ApiTags('Credits')
-@Controller('/api/v1/credits')
+@Controller('/api/v1/app')
 export class CreditsController {
   constructor(private readonly creditService: CreditService) {}
+  @ApiOkResponse({
+    description: 'Credit plan list',
+    type: CreditPlanResponseDto,
+    isArray: true,
+  })
+  @ApiQuery({
+    description: 'If staus then retur only Active status',
+    name: 'status',
+    required: false,
+  })
+  @UsePipes(new ValidationPipe({ transform: true }))
+  @Get(':appId/credits')
+  async getCredits(
+    @Param('appId') appId: string,
+    @Query() query: GetCreditsQueryDto,
+  ) {
+    const { status } = query;
+    return this.creditService.fetchCreditDetails(appId, status);
+  }
+
+  @ApiBearerAuth('Authorization')
+  @ApiOkResponse({
+    description: 'Credit plan activated successfully',
+    type: CreditPlanResponseDto,
+  })
+  @ApiParam({ name: 'creditId', description: 'Credit plan id' })
+  @Post(':appId/credits/:creditId/activate')
+  async activateCredit(
+    @Param('creditId') creditId: string,
+    @Param('appId') appId: string,
+  ) {
+    return this.creditService.activateCredit(creditId, appId);
+  }
+
   @ApiBearerAuth('Authorization')
   @ApiOkResponse({
     description: 'Credit granted successfully',
     type: CreditResponseDto,
   })
-  @Post('/:appId')
+  @UsePipes(new ValidationPipe({ transform: true }))
+  @Post(':appId/credits')
   async grantCredit(
     @Param('appId') appId: string,
     @Body() creditRequestDto: CreditRequestDto,
@@ -43,6 +82,7 @@ export class CreditsController {
       appId,
       creditRequestDto,
       req.user.userId,
+      CreditSourceEnum.MANUAL_RECHARGE,
     );
   }
 }