+++ src/credits/credits.module.ts
@@ -6,7 +6,7 @@ import {
   RequestMethod,
 } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { AuthZCredits, AuthZCreditsSchema } from './schemas/authz.schema';
+import { CreditPlan, CreditsSchema } from './schemas/credit.schema';
 import { CreditService } from './services/credits.service';
 import { CreditsController } from './controllers/credits.controller';
 import { JWTAuthorizeMiddleware } from 'src/utils/middleware/jwt-authorization.middleware';
@@ -22,12 +22,13 @@ import { HidWalletModule } from 'src/hid-wallet/hid-wallet.module';
 import { AppAuthModule } from 'src/app-auth/app-auth.module';
 import { RateLimitMiddleware } from 'src/utils/middleware/rate-limit.middleware';
 import { SuperAdminMiddleware } from 'src/utils/middleware/super-admin.middleware';
+import { CreditRepository } from './repositories/credit.repository';
 
 @Module({
   imports: [
     UserModule,
     MongooseModule.forFeature([
-      { name: AuthZCredits.name, schema: AuthZCreditsSchema },
+      { name: CreditPlan.name, schema: CreditsSchema },
     ]),
     MongooseModule.forFeature([
       { name: AdminPeople.name, schema: AdminPeopleSchema },
@@ -38,16 +39,28 @@ import { SuperAdminMiddleware } from 'src/utils/middleware/super-admin.middlewar
     HidWalletModule,
   ],
   controllers: [CreditsController],
-  providers: [AdminPeopleRepository, CreditService],
+  providers: [AdminPeopleRepository, CreditRepository, CreditService],
   exports: [CreditService],
 })
 export class CreditModule implements NestModule {
   configure(consumer: MiddlewareConsumer) {
-    consumer.apply(JWTAuthorizeMiddleware).forRoutes(CreditsController);
-    consumer.apply(JWTAccessAccountMiddleware).forRoutes(CreditsController);
+    consumer
+      .apply(JWTAuthorizeMiddleware)
+      .exclude({ path: 'api/v1/app/:appId/credits', method: RequestMethod.GET })
+      .forRoutes(CreditsController);
+    consumer
+      .apply(JWTAccessAccountMiddleware)
+      .exclude({ path: 'api/v1/app/:appId/credits', method: RequestMethod.GET })
+      .forRoutes(CreditsController);
     consumer
       .apply(SuperAdminMiddleware)
-      .exclude({ path: 'api/v1/credits/app', method: RequestMethod.GET })
+      .exclude(
+        { path: 'api/v1/app/:appId/credits', method: RequestMethod.GET },
+        {
+          path: 'api/v1/app/:appId/credits/:creditId/activate',
+          method: RequestMethod.POST,
+        },
+      )
       .forRoutes(CreditsController);
     consumer.apply(RateLimitMiddleware).forRoutes(CreditsController);
   }